cloud_requirements.txt
# Cloud-optimized requirements for the budget dashboard.
# Replaces dashboard_requirements.txt for Docker builds — strips macOS-only
# packages (appnope, etc.) and Jupyter tooling that is unused in the container.

# Core data
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
python-dateutil>=2.8.2
pytz>=2021.1
six>=1.16.0

# Dashboard
streamlit>=1.28.0,<2.0.0
altair>=4.2.0,<5.0.0
pydeck>=0.7.0

# Financial integrations
plaid-python>=8.6.0
requests>=2.26.0
urllib3>=1.26.7

# Config / env
python-dotenv>=0.19.0
toml>=0.10.2

# Optional: local ML classification
# Uncomment the lines below and rebuild if you want use_local_categorization=true.
# Note: adds ~3 GB to the image and significantly increases cold-start time.
# torch>=1.9.0
# transformers>=4.12.0
